Added carts MVC layers, tested curl scripts for Auth and Carts resource
1) Carts controller presets user authentication for all CRUD actions so
that each user can only make amendments to his/cart items and view his/her past orders.
2) Carts schema pastOrder field is of type array and will contain an array of key-value pairs of products
that the user wants to purchase.
3) Carts schema contains orderTotal field. This value will be sent to Stripe API along with confidential credit
card information for payment authentication.
4) Have to develop MVC layer for interaction with Stripe so that pastOrder array with validated item can be passed back as JSON on
success/item can be deleted from array on fail.
