+++ app/controllers/carts.js
@@ -1,0 +1,52 @@
+'use strict'
+
+const controller = require('lib/wiring/controller')
+const models = require('app/models')
+const Cart = models.cart
+
+const authenticate = require('./concerns/authenticate')
+const setUser = require('./concerns/set-current-user')
+const setModel = require('./concerns/set-mongoose-model')
+
+const index = (req, res, next) => {
+  // console.log('user is', this.setUser())
+  // console.log('req is', req)
+  Cart.find()
+    .then(carts => res.json({
+      carts: carts.map((e) =>
+      e.toJSON({user: req.user}))
+    }))
+    .catch(next)
+}
+
+const show = (req, res) => {
+  res.json({
+    cart: req.cart.toJSON({ user: req.user })
+  })
+}
+
+const create = (req, res, next) => {
+  // debugger
+  // console.log(req.body.carts)
+  const cart = Object.assign(req.body.cart, {
+    _owner: req.user._id
+  })
+  Cart.create(cart)
+    .then(cart =>
+      res.status(201)
+        .json({
+          cart: cart.toJSON({ user: req.user })
+        }))
+    .catch(next)
+}
+
+module.exports = controller({
+  index,
+  show,
+  create
+}, { before: [
+  // { method: setUser, only: ['index', 'show'] },
+  { method: authenticate, only: ['index', 'show', 'create', 'update', 'destroy'] },
+  { method: setModel(Cart), only: ['show'] },
+  { method: setModel(Cart, { forUser: true }), only: ['update', 'destroy'] }
+] })
