+++ package.json
@@ -15,8 +15,8 @@
     "cors": "^2.8.3",
     "debug": "^2.6.6",
     "dotenv": "^4.0.0",
-    "express": "^4.15.2",
-    "mongoose": "^4.13.11",
+    "express": "^4.16.2",
+    "mongoose": "^5.0.5",
     "mongoose-unique-validator": "^1.0.5",
     "morgan": "^1.8.1",
     "multer": "^1.3.0",
