+++ scripts/carts/destroy.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+API="http://localhost:4741"
+URL_PATH="/carts"
+
+curl "${API}${URL_PATH}/${ID}" \
+  --include \
+  --request DELETE \
+  --header "Authorization: Token token=${TOKEN}"
+
+echo
