+++ scripts/carts/show.sh
@@ -1,0 +1,11 @@
+#!/bin/sh
+
+API="http://localhost:4741"
+URL_PATH="/carts"
+
+curl "${API}${URL_PATH}/${ID}" \
+  --include \
+  --request GET \
+  --header "Authorization: Token token=$TOKEN"
+
+echo
