+++ scripts/carts/update.sh
@@ -1,0 +1,18 @@
+#!/bin/bash
+
+API="http://localhost:4741"
+URL_PATH="/carts"
+
+curl "${API}${URL_PATH}/${ID}" \
+  --include \
+  --request PATCH \
+  --header "Content-Type: application/json" \
+  --header "Authorization: Token token=${TOKEN}" \
+  --data '{
+    "cart": {
+      "pastOrder": "'"${PAST_ORDER}"'",
+      "orderTotal": "'"${TOTAL}"'"
+    }
+  }'
+
+echo
